pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ergos"
version = "0.1.0"
description = "Local-first voice assistant with STT, LLM, and TTS pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Ergos Contributors"}
]
keywords = ["voice-assistant", "stt", "tts", "llm", "local-first"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "torch>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "faster-whisper>=1.0",
    "llama-cpp-python>=0.2",
    "kokoro-onnx>=0.4",
]

[project.optional-dependencies]
csm = [
    "transformers>=4.52.1",
    "torchaudio>=2.0",
    "accelerate>=0.20",
]
orpheus = [
    "orpheus-cpp>=0.0.3",
]
cloud = [
    "openai>=1.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
ergos = "ergos.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
